fix(cron): persist lastTickMinute across scheduler stop/start cycle
The previous fix read scheduler?.getLastTickMinute() in startScheduler(),
but stopScheduler() already set scheduler = null — so the value was
always undefined and jobs could re-fire within the same minute.

Save lastTickMinute to a module-level variable before nulling the
scheduler ref so it survives the stop → start gap.

Also split scheduler.test.ts (595 LOC) into scheduler.test.ts (399)
and scheduler-reminders.test.ts (281) to stay within the 500 LOC limit.
