+++ packages/pi-cron-extension/extension/__tests__/scheduler.test.ts
@@ -397,199 +397,3 @@ describe('CronScheduler lifecycle', () => {
     expect(scheduler.isRunning()).toBe(false);
   });
 });
-
-// ── Restart with lastTickMinute (duplicate prevention fix) ───────
-
-describe('CronScheduler restart dedup', () => {
-  it('does not re-fire jobs when restarted with same lastTickMinute', async () => {
-    mockMatchesCron.mockReturnValue(true);
-    vi.setSystemTime(new Date('2025-06-15T09:00:00'));
-
-    const scheduler1 = new CronScheduler();
-    scheduler1.start([makeJob()], '/test');
-    await vi.advanceTimersByTimeAsync(0);
-    expect(mockRunTransientSession).toHaveBeenCalledTimes(1);
-
-    const prevMinute = scheduler1.getLastTickMinute();
-    scheduler1.stop();
-    mockRunTransientSession.mockClear();
-
-    // Restart within the same minute, passing lastTickMinute
-    const scheduler2 = new CronScheduler();
-    scheduler2.start([makeJob()], '/test', [], { lastTickMinute: prevMinute });
-    await vi.advanceTimersByTimeAsync(0);
-
-    // Should NOT fire again in the same minute
-    expect(mockRunTransientSession).not.toHaveBeenCalled();
-
-    scheduler2.stop();
-  });
-
-  it('fires jobs in a new minute after restart', async () => {
-    mockMatchesCron.mockReturnValue(true);
-    vi.setSystemTime(new Date('2025-06-15T09:00:00'));
-
-    const scheduler1 = new CronScheduler();
-    scheduler1.start([makeJob()], '/test');
-    await vi.advanceTimersByTimeAsync(0);
-    expect(mockRunTransientSession).toHaveBeenCalledTimes(1);
-
-    const prevMinute = scheduler1.getLastTickMinute();
-    scheduler1.stop();
-    mockRunTransientSession.mockClear();
-
-    // Advance to the next minute
-    vi.setSystemTime(new Date('2025-06-15T09:01:00'));
-
-    const scheduler2 = new CronScheduler();
-    scheduler2.start([makeJob()], '/test', [], { lastTickMinute: prevMinute });
-    await vi.advanceTimersByTimeAsync(0);
-
-    // Should fire in the new minute
-    expect(mockRunTransientSession).toHaveBeenCalledTimes(1);
-
-    scheduler2.stop();
-  });
-
-  it('fires jobs normally without lastTickMinute option', async () => {
-    mockMatchesCron.mockReturnValue(true);
-    vi.setSystemTime(new Date('2025-06-15T09:00:00'));
-
-    const scheduler = new CronScheduler();
-    scheduler.start([makeJob()], '/test');
-    await vi.advanceTimersByTimeAsync(0);
-
-    expect(mockRunTransientSession).toHaveBeenCalledTimes(1);
-
-    scheduler.stop();
-  });
-});
-
-// ── Reminder tick tests ──────────────────────────────────────────
-
-describe('CronScheduler reminder ticks', () => {
-  it('fires onReminderFire when shouldFire returns true', async () => {
-    mockShouldFire.mockReturnValue(true);
-
-    const onReminderFire = vi.fn();
-    const onReminderUpdate = vi.fn();
-    const scheduler = new CronScheduler({ onReminderFire, onReminderUpdate });
-    scheduler.start([], '/test', [makeReminder()]);
-
-    await vi.advanceTimersByTimeAsync(0);
-
-    expect(onReminderFire).toHaveBeenCalledWith(
-      expect.objectContaining({ id: 'rem-1', title: 'Test Reminder' }),
-    );
-  });
-
-  it('calls onReminderUpdate with post-fire state', async () => {
-    mockShouldFire.mockReturnValue(true);
-
-    const onReminderUpdate = vi.fn();
-    const scheduler = new CronScheduler({ onReminderUpdate });
-    scheduler.start([], '/test', [makeReminder()]);
-
-    await vi.advanceTimersByTimeAsync(0);
-
-    expect(onReminderUpdate).toHaveBeenCalledWith(
-      expect.objectContaining({ status: 'completed' }),
-    );
-  });
-
-  it('updates in-memory reminder to prevent re-firing on next tick', async () => {
-    let callCount = 0;
-    mockShouldFire.mockImplementation(() => {
-      callCount++;
-      // Only fire on the first call
-      return callCount === 1;
-    });
-
-    const onReminderFire = vi.fn();
-    const scheduler = new CronScheduler({ onReminderFire });
-    scheduler.start([], '/test', [makeReminder()]);
-
-    // First tick
-    await vi.advanceTimersByTimeAsync(0);
-    expect(onReminderFire).toHaveBeenCalledTimes(1);
-
-    // Second tick (30s later) — shouldFire returns false
-    await vi.advanceTimersByTimeAsync(30_000);
-    expect(onReminderFire).toHaveBeenCalledTimes(1);
-
-    scheduler.stop();
-  });
-
-  it('does not fire reminders when list is empty', async () => {
-    // Reset all mocks to ensure clean state
-    mockShouldFire.mockReset();
-    mockShouldFire.mockReturnValue(false);
-
-    const onReminderFire = vi.fn();
-    const scheduler = new CronScheduler({ onReminderFire });
-    scheduler.start([], '/test', []);
-
-    // First tick + one interval tick
-    await vi.advanceTimersByTimeAsync(30_000);
-
-    expect(onReminderFire).not.toHaveBeenCalled();
-    // shouldFire should never be called because reminders list is empty
-    // (the tickReminders method early-returns for empty list)
-    expect(mockShouldFire).not.toHaveBeenCalled();
-
-    scheduler.stop();
-  });
-
-  it('fires multiple reminders in the same tick', async () => {
-    mockShouldFire.mockReturnValue(true);
-
-    const onReminderFire = vi.fn();
-    const scheduler = new CronScheduler({ onReminderFire });
-    scheduler.start([], '/test', [
-      makeReminder({ id: 'a', title: 'First' }),
-      makeReminder({ id: 'b', title: 'Second' }),
-    ]);
-
-    await vi.advanceTimersByTimeAsync(0);
-
-    expect(onReminderFire).toHaveBeenCalledTimes(2);
-    const ids = onReminderFire.mock.calls.map((c: any[]) => c[0].id);
-    expect(ids).toContain('a');
-    expect(ids).toContain('b');
-
-    scheduler.stop();
-  });
-
-  it('reminders are checked every tick (not limited to once per minute like jobs)', async () => {
-    mockShouldFire.mockReset();
-    mockStatusAfterFire.mockReset();
-
-    // shouldFire always returns true (simulating recurring reminder)
-    mockShouldFire.mockReturnValue(true);
-    // statusAfterFire keeps it active (recurring behavior)
-    mockStatusAfterFire.mockImplementation((r: any) => ({
-      ...r,
-      status: 'active',
-      lastFiredAt: new Date().toISOString(),
-    }));
-
-    vi.setSystemTime(new Date('2025-06-15T09:00:00'));
-
-    const onReminderFire = vi.fn();
-    const scheduler = new CronScheduler({ onReminderFire });
-    scheduler.start([], '/test', [makeReminder({ type: 'recurring', schedule: '* * * * *' })]);
-
-    // First tick (immediate in start())
-    await vi.advanceTimersByTimeAsync(0);
-    const firstTickFireCount = onReminderFire.mock.calls.length;
-    expect(firstTickFireCount).toBeGreaterThanOrEqual(1);
-
-    // Second tick (30s later, same minute) — reminders are checked every tick
-    const shouldFireCallsBefore = mockShouldFire.mock.calls.length;
-    await vi.advanceTimersByTimeAsync(30_000);
-    // shouldFire should have been called again on the second tick
-    expect(mockShouldFire.mock.calls.length).toBeGreaterThan(shouldFireCallsBefore);
-
-    scheduler.stop();
-  });
-});
